fixed some major issues

+++ main_app/routers/users.py
@@ -62,6 +62,9 @@ def edit_user(user_in: UserUpdate, current_user: User = Depends(get_current_user
 
 @router.delete("/{user_id}", response_model=Message)
 def remove_user(user_id: int, current_user: User = Depends(get_admin_user), db: Session = Depends(get_db)):
+    if current_user.id == user_id:
+        raise AppException(status.HTTP_401_UNAUTHORIZED, "Cannot delete admin account")
+
     return delete_user(db, user_id)
 
 @router.get("/me/role", response_model=UserRoleResponse)
